% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Coefficeients of the estimated regression modell}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{}
}
\value{

}
\description{
Coefficeients of the estimated regression modell
}
