% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear regression with QR}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{A forumula describin the model}

\item{data}{A data.frame containing the data}
}
\value{
Returns a linear regression model
}
\description{
Linear regression with QR
}
\examples{
linreg(Sepal.Length ~ Species, data = iris)

}
