% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Show two scatter plots.The first one fitted values against Residuals.
The second scatter plot show square root of standardized residuals against fitted values}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{is a linreg object}

\item{...}{extra stuff}
}
\value{
Produces two scatter plots, one for residuals against fitted values and standardized residuals against fits
}
\description{
Show two scatter plots.The first one fitted values against Residuals.
The second scatter plot show square root of standardized residuals against fitted values
}
\examples{
model <- linreg(Sepal.Length ~ Species, data = iris)
plot(model)
}
