% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred.linreg}
\alias{pred.linreg}
\title{Pred}
\usage{
\method{pred}{linreg}(x)
}
\arguments{
\item{x}{is a linreg object}
}
\value{
Returns the fitted values of the regression model on the training data
}
\description{
Pred
}
\examples{
model <- linreg(Sepal.Length ~ Species, data = iris)
pred(model)

}
