% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_pred.R
\name{pred}
\alias{pred}
\title{Creates generic function for pred}
\usage{
pred(x)
}
\arguments{
\item{x}{blablab}
}
\value{
A generic function called pred
}
\description{
Creates generic function for pred
}
