% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.linreg}
\alias{print.linreg}
\title{Print method for linreg classes}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{is a linreg object}

\item{...}{extra stuff}
}
\value{
Returns the coefficients and the formula of the fitted model.
}
\description{
Print method for linreg classes
}
\examples{
model <- linreg(Sepal.Length ~ Species, data = iris)
print(model)
}
