% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{resid.linreg}
\alias{resid.linreg}
\title{Residuals of a linereg class model}
\usage{
\method{resid}{linreg}(object)
}
\arguments{
\item{object}{}
}
\value{

}
\description{
Residuals of a linereg class model
}
