% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{residuals.linreg}
\alias{residuals.linreg}
\title{Residuals of a linereg class model}
\usage{
\method{residuals}{linreg}(object, ...)
}
\arguments{
\item{object}{is a linreg object}

\item{...}{extra stuff}
}
\value{
The function returns a vector with residuals of the fitted model
}
\description{
Residuals of a linereg class model
}
\examples{
model <- linreg(Sepal.Length ~ Species, data = iris)
resid(model)
}
