% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_linkoping.R
\name{theme_linkoping}
\alias{theme_linkoping}
\title{ggplot theme for Linköping university}
\usage{
theme_linkoping()
}
\value{
Adds linkoping theme to ggplot2's plot
}
\description{
ggplot theme for Linköping university
}
