% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Show two scatter plots.The first one fitted values against Residuals.
The second scatter plot show square root of standardized residuals against fitted values}
\usage{
\method{plot}{linreg}(object)
}
\arguments{
\item{object}{}
}
\value{

}
\description{
Show two scatter plots.The first one fitted values against Residuals.
The second scatter plot show square root of standardized residuals against fitted values
}
