% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{residuals.linreg}
\alias{residuals.linreg}
\title{Residuals of a linereg class model}
\usage{
\method{residuals}{linreg}(object, ...)
}
\arguments{
\item{object}{}
}
\value{
Vector with residuals
}
\description{
Residuals of a linereg class model
}
