% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Summary printout of the fitted regerssionmodel}
\usage{
\method{summary}{linreg}(object)
}
\arguments{
\item{object}{}
}
\value{

}
\description{
Summary printout of the fitted regerssionmodel
}
